+++ src/pdfjsLoader.ts
@@ -5,17 +5,13 @@ import { propsToPdfDocInitParams } from './propsToPdfDocInitParams.js';
 
 let pdfjsLib: typeof PdfjsModule | undefined;
 
-export async function getPdfDocument(
-    pdfFileBuffer: Uint8Array | ArrayBufferLike,
-    opts: NormalizedPdfToPngOptions,
-): Promise<PDFDocumentProxy> {
+export async function getPdfDocument(pdfFileBuffer: Uint8Array, opts: NormalizedPdfToPngOptions): Promise<PDFDocumentProxy> {
     pdfjsLib ??= await import('pdfjs-dist/legacy/build/pdf.mjs');
     const { getDocument } = pdfjsLib;
     const documentInitParameters = propsToPdfDocInitParams(opts);
-    const data: Uint8Array = pdfFileBuffer instanceof Uint8Array ? pdfFileBuffer : new Uint8Array(pdfFileBuffer);
     const task: PDFDocumentLoadingTask = getDocument({
         ...documentInitParameters,
-        data,
+        data: pdfFileBuffer,
     });
 
     try {